Strip out Carli::Node_Fringe_Ranged; Strip out Range/Lines tracking for nodes; Begin respecialization work;

+++ src/carli/rete/rete_agent.h
@@ -53,7 +53,7 @@ namespace Rete {
     std::unordered_map<std::string, Rete_Action_Ptr> rules;
     WME_Set working_memory;
     Agenda agenda;
-    int visitor_value = 0;
+    intptr_t visitor_value = 0;
   };
 
 }
